requirements.txt
python-dotenv>=0.21.0
motor>=3.1.1
SQLAlchemy>=1.4.0
psycopg2>=2.9.0
Telethon>=1.41.2
pyromod>=1.5.0
TgCrypto>=1.0.0
pyrogram>=2.0.106
flask>=2.3.0
flask-restful>=0.3.0
requests>=2.31.0
pytz>=2023.3
psutil>=5.9.5
platformdirs>=2.6.2
pymongo
